requirements.txt
astroid==3.0.3
attrs==23.2.0
certifi==2024.2.2
dill==0.3.8
distlib==0.3.8
exceptiongroup==1.2.0
filelock==3.13.1
h11==0.14.0
idna==3.6
isort==5.13.2
mccabe==0.7.0
outcome==1.3.0.post0
platformdirs==4.2.0
pylint==3.0.3
PySocks==1.7.1
python-dotenv==1.0.1
selenium==4.17.2
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
tomlkit==0.12.3
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.2.0
virtualenv==20.25.0
wsproto==1.2.0
